+++ Common/WebStore.Domain/ViewModels/BrandViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
     public class BrandViewModel : INamedEntity, IOrderedEntity
     {
