+++ Common/WebStore.Domain/ViewModels/Orders/UserOrderViewModel.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace WebStore.ViewModels.Orders
+﻿namespace WebStore.Domain.ViewModels.Orders
 {
     public class UserOrderViewModel
     {
