+++ UI/WebStore/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,7 +1,4 @@
-﻿@using WebStore.Models
-@using WebStore.ViewModels
-@using WebStore.ViewModels.Identity
-@using WebStore.Domain.Entities
+﻿@using WebStore.Domain.Entities
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, WebStore
