+++ UI/WebStore/Data/TestData.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using WebStore.Domain.Entities;
-using WebStore.Models;
+using WebStore.Domain.Models;
 
 namespace WebStore.Data
 {
