+++ UI/WebStore/Infrastructure/Interfaces/IEmployeesData.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using WebStore.Models;
+using WebStore.Domain.Models;
 
 namespace WebStore.Infrastructure.Interfaces
 {
