+++ UI/WebStore/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model RegisterUserViewModel
+﻿@model WebStore.Domain.ViewModels.Identity.RegisterUserViewModel
 @{
     ViewData["Title"] = "Регистрация нового пользователя";
 }
