+++ UI/WebStore/Views/Cart/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model CartOrderViewModel
+﻿@model WebStore.Domain.ViewModels.CartOrderViewModel
 @{
     ViewData["Title"] = "Корзина";
 }
