+++ UI/WebStore/Views/Catalog/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProductViewModel
+﻿@model WebStore.Domain.ViewModels.ProductViewModel
 @{
     ViewData["Title"] = $"Товар {Model.Name}";
 }
