+++ UI/WebStore/Views/Catalog/Shop.cshtml
@@ -1,4 +1,4 @@
-﻿@model CatalogViewModel
+﻿@model WebStore.Domain.ViewModels.CatalogViewModel
 
 <section id="advertisement">
     <div class="container">
