+++ UI/WebStore/Views/Employees/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model EmployeeViewModel
+﻿@model WebStore.Domain.ViewModels.EmployeeViewModel
 
 @{
     ViewData["Title"] = "View";
