+++ UI/WebStore/Views/Employees/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model EmployeeViewModel
+﻿@model WebStore.Domain.ViewModels.EmployeeViewModel
 
 @{
     ViewData["Title"] = "Сотрудник #" + Model.Id;
