+++ UI/WebStore/Views/Employees/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model EmployeeViewModel
+﻿@model WebStore.Domain.ViewModels.EmployeeViewModel
 
 @{
     ViewData["Title"] = "Редактирование сотрудника id:" + Model.Id;
