+++ UI/WebStore/Views/Employees/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<EmployeeViewModel>
+﻿@model IEnumerable<WebStore.Domain.ViewModels.EmployeeViewModel>
 
 @{
     //ViewData["Title"] = "Сотрудники";
