+++ UI/WebStore/Views/Shared/Components/Sections/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<SectionViewModel>
+﻿@model IEnumerable<WebStore.Domain.ViewModels.SectionViewModel>
 
 <h2>Категории</h2>
 <div class="panel-group category-products" id="accordian">
