+++ UI/WebStore/Views/Shared/Partial/_FeaturesItems.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ProductViewModel>
+﻿@model IEnumerable<WebStore.Domain.ViewModels.ProductViewModel>
 
 <div class="features_items">
     <!--features_items-->
