+++ UI/WebStore/Views/_ViewImports.cshtml
@@ -1,6 +1,2 @@
-﻿@using WebStore.Models
-@using WebStore.ViewModels
-@using WebStore.ViewModels.Identity
-
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, WebStore
